+++ src/clj_depend/analyzer.clj
@@ -7,22 +7,21 @@
 
 (defn- violate?
   [config
-   {:keys [layer dependent-layer]}]
-  (let [accessed-by-layers (get-in config [:layers layer :accessed-by-layers])]
-    (not-any? (partial = dependent-layer) accessed-by-layers)))
+   {:keys [layer dependency-layer]}]
+  (when-let [accessed-by-layers (get-in config [:layers dependency-layer :accessed-by-layers])]
+    (not-any? (partial = layer) accessed-by-layers)))
 
-(defn- layer-and-namespace [config namespace dependent-namespace]
-  (let [layer (layer-by-namespace config namespace)
-        dependent-layer (layer-by-namespace config dependent-namespace)]
-    {:namespace           namespace
-     :dependent-namespace dependent-namespace
-     :layer               layer
-     :dependent-layer     dependent-layer}))
+(defn- layer-and-namespace [config namespace dependency-namespace]
+  (when-let [layer (layer-by-namespace config namespace)]
+    {:namespace            namespace
+     :layer                layer
+     :dependency-namespace dependency-namespace
+     :dependency-layer     (layer-by-namespace config dependency-namespace)}))
 
 (defn- violations
   [config dependency-graph namespace]
-  (let [dependent-namespaces (dependency/immediate-dependents dependency-graph namespace)]
-    (->> dependent-namespaces
+  (let [dependencies (dependency/immediate-dependencies dependency-graph namespace)]
+    (->> dependencies
          (map #(layer-and-namespace config namespace %))
          (filter #(violate? config %))
          not-empty)))
@@ -32,7 +31,7 @@
   [config namespaces]
   (let [dependency-graph (dependency/dependencies-graph namespaces)
         violations (flatten (keep #(violations config dependency-graph (:name %)) namespaces))]
-    (map (fn [{:keys [namespace dependent-namespace]}]
-           {:namespace dependent-namespace
-            :violation namespace})
+    (map (fn [{:keys [namespace dependency-namespace]}]
+           {:namespace namespace
+            :violation dependency-namespace})
          violations)))