+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.3.0 (2022-05-09)
+* Fixed analyzer to only consider immediate dependencies for each namespace.
+* Extracted lein plugin to another project (lein-clj-depend).
+
 ## 0.2.0 (2022-04-27)
 * Improve exit codes.